docs(site): Moved all doc/example code into site package.
44
