--- packages/archetype/src/__demo__/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width">
-    <title>Demo</title>
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-</html>
