+++ packages/hello-world/src/__demo__/examples/Default.js
@@ -15,9 +15,10 @@
  */
 
 /* @flow */
-import React from 'react';
 import HelloWorld from '../../HelloWorld';
 
-export default function Default() {
-  return <HelloWorld />;
-}
+export default {
+  title: 'Default',
+  component: HelloWorld,
+  source: `<HelloWorld />`
+};
