+++ packages/site/package.json
@@ -15,7 +15,7 @@
     "jest": "$npm_package_config_script_path/jest.sh",
     "lint": "$npm_package_config_script_path/lint.sh",
     "pack-ls": "$npm_package_config_script_path/pack-ls.sh",
-    "start": "TARGET=site $npm_package_config_script_path/start.sh",
+    "start": "TARGET=site webpack-dev-server --progress",
     "test": "$npm_package_config_script_path/test.sh"
   },
   "devDependencies": {
