+++ packages/style-utils/src/__demo__/examples/componentTheme.js
@@ -19,7 +19,7 @@ import React from 'react';
 import { ThemeProvider } from '../../index';
 import Sample from '../../Sample';
 
-export default function ComponentTheme() {
+function ComponentTheme() {
   return (
     <div>
       <ThemeProvider theme={{ Sample_color: 'mediumvioletred' }}>
@@ -28,3 +28,12 @@ export default function ComponentTheme() {
     </div>
   );
 }
+
+export default {
+  title: 'Component theme var override',
+  component: ComponentTheme,
+  description: 'This will affect all components inside this ThemeProvider that use this variable, which should only be Sample components due to the namespace.',
+  source: `<ThemeProvider theme={{ Sample_color: 'mediumvioletred' }}>
+<Sample />
+</ThemeProvider>`
+};
