+++ packages/style-utils/src/__demo__/examples/createThemedComponent.js
@@ -23,10 +23,21 @@ const MyThemedSample = createThemedComponent(Sample, {
   color_primary: 'mediumvioletred'
 });
 
-export default function CreateThemedComponent() {
+function CreateThemedComponent() {
   return (
     <div>
       <MyThemedSample />
     </div>
   );
 }
+
+export default {
+  title: 'Local theme var override via createThemedComponent',
+  component: CreateThemedComponent,
+  description: 'If you need to overwrite a theme variable, either global or component-specific, for a specific component.',
+  source: `const MyThemedSample = createThemedComponent(Sample, {
+color_primary: 'mediumvioletred'
+});
+
+<MyThemedSample />`
+};
