+++ packages/world/src/__demo__/index.js
@@ -15,14 +15,14 @@
  */
 
 /* @flow */
-import React from 'react';
-import { render } from 'react-dom';
-import { ThemeProvider } from '@mineral-ui/style-utils';
-import App from './App';
 
-render(
-  <ThemeProvider>
-    <App />
-  </ThemeProvider>,
-  document.getElementById('app')
-);
+import examples from './examples';
+
+export default [
+  {
+    description: 'A simple component that renders the string, "World". It is primarily used for testing project configuration and package distribution.',
+    examples,
+    slug: 'world',
+    title: 'World'
+  }
+];
