+++ scripts/package/start.sh
@@ -1,5 +1,4 @@
 #!/bin/bash -e
 
-TARGET="${TARGET:-demo}"
-
-TARGET=$TARGET webpack-dev-server --progress
+cd ../../
+npm start
